better generation

+++ examples/examples_1D/01_linearadvection/mcT_adv_data.py
@@ -1,6 +1,8 @@
 from typing import Tuple, NamedTuple
 import shutil, os, json, re
 import matplotlib.pyplot as plt
+import jax
+from jax import vmap, pmap, lax
 from jaxfluids import InputReader, Initializer, SimulationManager
 from jaxfluids.post_process import load_data, create_lineplot
 import jax.numpy as jnp
@@ -27,6 +29,9 @@ class Data():
         self.sims = os.listdir(self.save_path)
         self.sims.sort()
 
+    def size(self) -> int:
+        return len(self.sims)
+
     def next_sim(self) -> Sim:
 
         sim_path = self.sims.pop()
@@ -43,19 +48,34 @@ class Data():
 
         return Sim(domain, case_setup, num_setup)
 
-    def generate(self):
-        shutil.rmtree(self.save_path)
+    def _generate(self, *args):
+        case_setup = setup.cases.next()
+        num_setup = setup.numerical
 
-        for ii in range(setup.cases.size()):
-            case_setup = setup.cases.next()
-            num_setup = setup.numerical
+        input_reader = InputReader(case_setup, num_setup)
+        initializer  = Initializer(input_reader)
+        sim_manager  = SimulationManager(input_reader)
 
-            input_reader = InputReader(case_setup, num_setup)
-            initializer  = Initializer(input_reader)
-            sim_manager  = SimulationManager(input_reader)
+        buffer_dictionary = initializer.initialization()
+        sim_manager.simulate(buffer_dictionary)
 
-            buffer_dictionary = initializer.initialization()
-            sim_manager.simulate(buffer_dictionary)
+    def generate(self, reset=False):
+        if reset:
+            shutil.rmtree(self.save_path)
+            
+        self.check_sims()
+        if self.size() < setup.cases.size():
+            for _ in range(self.size()):
+                _ = setup.cases.next()
+            n_to_gen = jnp.arange(setup.cases.size() - self.size())
+            
+            # batch for parallel and vectorized generation
+            n_dev = jax.local_device_count()
+            batch_idx = jnp.linspace(0,n_to_gen.shape[0],n_dev+1)
+            n_to_gen_par = jnp.array([n_to_gen[ii:jj,...] for ii,jj in zip(batch_idx[:-1],batch_idx[1:])])
+
+            pmap(vmap(self._generate, in_axes=(0,)), in_axes=(0,))(n_to_gen_par)
+            
 
 data = Data(setup.save_path)
